website/app/requirements.txt
Flask
gunicorn
pyyaml
requests
prometheus-flask-exporter